Refactoring is done
