+++ hw2_4/hw2_4/Program.cs
@@ -1,15 +1,19 @@
 ﻿using System;
-using NamespaceCalculator;
-using NamespaceStack;
+using Stacks;
+using Calculator;
 
 namespace hw2_4
 {
 	class Program
 	{
+		/// <summary>
+		/// The entry point of the program, where the program control starts and ends.
+		/// </summary>
+		/// <param name="args">The command-line arguments.</param>
 		public static void Main (string[] args)
 		{
 			ArrayStack myStack = new ArrayStack ();
-			Calculator myCalculator = new Calculator(myStack);
+			StackCalculator myCalculator = new StackCalculator(myStack);
 			myCalculator.Push (3);
 			myCalculator.Push (2);
 			myCalculator.Push (1);
@@ -18,7 +22,7 @@ namespace hw2_4
 			Console.WriteLine (myCalculator.Result ());
 
 			ListStack myStack2 = new ListStack ();
-			Calculator myCalculator2 = new Calculator(myStack2);
+			StackCalculator myCalculator2 = new StackCalculator(myStack2);
 			myCalculator2.Push (1);
 			myCalculator2.Push (2);
 			myCalculator2.Push (3);
